optimisation de code + suppression des console.log
